Cleaning up CUDA code base
- removed superfluous headers
- fixed build system
- tabs --> spaces
- renamed server/*.cpp to make VS compile things in parallel
