+++ .gitignore
@@ -20,3 +20,4 @@
 
 # images
 *.png
+build
