+++ cuda.hpp
@@ -1,5 +1,5 @@
 // Copyright (c)    2013 Martin Stumpf
-//					2015 Patrick Diehl
+//                  2015 Patrick Diehl
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
@@ -8,7 +8,7 @@
     #include "cuda/device.hpp"
     #include "cuda/buffer.hpp"
     #include "cuda/kernel.hpp"
-	#include "cuda/program.hpp"
+    #include "cuda/program.hpp"
     //#include "cuda/std.hpp"
-	#include "cuda/get_devices.hpp"
+    #include "cuda/get_devices.hpp"
 #endif
