+++ cuda/device.cpp
@@ -1,25 +1,17 @@
-// Copyright (c)		2013 Damond Howard
-//						2015 Patrick Diehl
+// Copyright (c)    2013 Damond Howard
+//                  2015 Patrick Diehl
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt
 
-//#include <hpx/hpx.hpp>
-//#include <hpx/runtime/components/component_factory.hpp>
-//#include <hpx/util/portable_binary_iarchive.hpp>
-//#include <hpx/util/portable_binary_oarchive.hpp>
+#include <hpx/hpx.hpp>
 
-//#include <boost/serialization/version.hpp>
-//#include <boost/serialization/export.hpp>
-
-#include "device.hpp"
-
-//HPX_REGISTER_COMPONENT_MODULE();
+#include "cuda/device.hpp"
 
 typedef hpx::components::managed_component<
-	hpx::cuda::server::device
-	> cuda_device_type;
+    hpx::cuda::server::device
+    > cuda_device_type;
 
-HPX_REGISTER_MINIMAL_COMPONENT_FACTORY(cuda_device_type,device);
+HPX_REGISTER_MINIMAL_COMPONENT_FACTORY(cuda_device_type, cuda_device);
 
 HPX_REGISTER_ACTION(
     cuda_device_type::wrapped_type::get_cuda_info_action,
@@ -46,8 +38,14 @@ HPX_REGISTER_ACTION(
     cuda_device_type::wrapped_type::mem_cpy_d_to_h_action,
     cuda_device_mem_cpy_d_to_h_action);
 HPX_REGISTER_ACTION(
-    cuda_device_type::wrapped_type::launch_kernel_action, 
+    cuda_device_type::wrapped_type::launch_kernel_action,
     cuda_device_launch_kernel_action);
 HPX_REGISTER_ACTION(
     cuda_device_type::wrapped_type::free_action,
     cuda_device_free_action);
+HPX_REGISTER_ACTION(
+    cuda_device_type::wrapped_type::create_program_with_source_action,
+    create_program_with_source_action);
+HPX_REGISTER_ACTION(
+    cuda_device_type::wrapped_type::create_buffer_action,
+    create_buffer_action);
