+++ cuda/fwd_declarations.hpp
@@ -8,22 +8,22 @@
 
 namespace hpx
 {
-	namespace cuda
-	{
-		class device;
-		class buffer;
-		class kernel;
-		class program;
-		class event;
-		namespace server
-		{
-			class device;
-			class buffer;
-			class kernel;
-			class program;
-			class event;
-		}
-	}
+    namespace cuda
+    {
+        class device;
+        class buffer;
+        class kernel;
+        class program;
+        class event;
+        namespace server
+        {
+            class device;
+            class buffer;
+            class kernel;
+            class program;
+            class event;
+        }
+    }
 }
 
 #endif
