+++ cuda/get_devices.hpp
@@ -6,19 +6,17 @@
 #ifndef HPX_CUDA_GET_DEVICES_HPP_
 #define HPX_CUDA_GET_DEVICES_HPP_
 
-#include <hpx/hpx.hpp>
-#include <hpx/config.hpp>
-#include <hpx/runtime/actions/component_action.hpp>
-#include <hpx/runtime/applier/apply.hpp>
-#include <hpx/include/async.hpp>
+#include <hpx/include/lcos.hpp>
 
-#include "export_definitions.hpp"
+#include "cuda/fwd_declarations.hpp"
+#include "cuda/export_definitions.hpp"
+#include "cuda/server/get_devices.hpp"
+#include "cuda.hpp"
 
 #include <cuda.h>
-#include "server/get_devices.hpp"
 
-#include "fwd_declarations.hpp"
-#include "../cuda.hpp"
+#include <vector>
+
 
 ////////////////////////////////////////////////////////////////
 namespace hpx {
@@ -45,9 +43,9 @@ namespace cuda {
  *                            Recommended value is "OpenCL 1.1".
  * @return A list of suitable OpenCL devices on target node
  */
-HPX_CUDA_EXPORT hpx::lcos::future<std::vector<device>>
+HPX_CUDA_EXPORT hpx::future<std::vector<device> >
 get_devices( hpx::naming::id_type node_id,
-		int major = 1 , int minor = 0 );
+        int major = 1 , int minor = 0 );
 /**
  * @brief Fetches a list of all accelerator devices present in the current
  *        hpx environment.
@@ -69,7 +67,7 @@ get_devices( hpx::naming::id_type node_id,
  *                            Recommended value is "OpenCL 1.1".
  * @return A list of suitable OpenCL devices
  */
-HPX_CUDA_EXPORT hpx::lcos::future<std::vector<device>>
+HPX_CUDA_EXPORT hpx::future<std::vector<device>>
 get_all_devices( int major = 1, int minor = 0 );
 /**
  * @brief Fetches a list of local accelerator devices present in the current
@@ -92,7 +90,7 @@ get_all_devices( int major = 1, int minor = 0 );
  *                            Recommended value is "OpenCL 1.1".
  * @return A list of suitable OpenCL devices
  */
-HPX_CUDA_EXPORT hpx::lcos::future<std::vector<device>>
+HPX_CUDA_EXPORT hpx::future<std::vector<device>>
 get_local_devices(int major = 1, int minor = 0);
 /**
  * @brief Fetches a list of remote accelerator devices present in the current
@@ -115,9 +113,9 @@ get_local_devices(int major = 1, int minor = 0);
  *                            Recommended value is "OpenCL 1.1".
  * @return A list of suitable OpenCL devices
  */
-HPX_CUDA_EXPORT hpx::lcos::future<std::vector<device>>
+HPX_CUDA_EXPORT hpx::future<std::vector<device>>
 get_remote_devices(
-		int major = 1, int minor = 0 );
+        int major = 1, int minor = 0 );
 }
 }
