+++ cuda/program.cpp
@@ -1,29 +1,21 @@
-// Copyright (c)		2013 Damond Howard
-//						2015 Patrick Diehl
+// Copyright (c)    2013 Damond Howard
+//                  2015 Patrick Diehl
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt
 
-//#include <hpx/hpx.hpp>
-//#include <hpx/runtime/components/component_factory.hpp>
-//#include <hpx/util/portable_binary_iarchive.hpp>
-//#include <hpx/util/portable_binary_oarchive.hpp>
+#include <hpx/hpx.hpp>
 
-//#include <boost/serialization/version.hpp>
-//#include <boost/serialization/export.hpp>
-
-#include "program.hpp"
-
-//HPX_REGISTER_COMPONENT_MODULE();
+#include "cuda/program.hpp"
 
 typedef hpx::components::managed_component<
     hpx::cuda::server::program>
     cuda_program_type;
 
-HPX_REGISTER_MINIMAL_COMPONENT_FACTORY(cuda_program_type,program);
+HPX_REGISTER_MINIMAL_COMPONENT_FACTORY(cuda_program_type, cuda_program);
 
 HPX_REGISTER_ACTION(cuda_program_type::wrapped_type::build_action,
-	cuda_program_build_action);
-HPX_REGISTER_ACTION(cuda_program_type::wrapped_type::set_source_action, 
-	cuda_program_set_source_action);
+    cuda_program_build_action);
+HPX_REGISTER_ACTION(cuda_program_type::wrapped_type::set_source_action,
+    cuda_program_set_source_action);
 HPX_REGISTER_ACTION(cuda_program_type::wrapped_type::create_kernel_action,
-	cuda_program_create_kernel_action);
+    cuda_program_create_kernel_action);
