+++ cuda/program.hpp
@@ -11,45 +11,46 @@
 
 namespace hpx {
 namespace cuda {
+
 class program: public hpx::components::client_base<program, server::program> {
-	typedef hpx::components::client_base<program, server::program> base_type;
+    typedef hpx::components::client_base<program, server::program> base_type;
 
 public:
-	program() {
-	}
+    program() {
+    }
 
-	program(hpx::future<hpx::naming::id_type> && gid) :
-			base_type(std::move(gid)) {
-	}
+    program(hpx::future<hpx::naming::id_type> && gid) :
+            base_type(std::move(gid)) {
+    }
 
-	hpx::lcos::future<void> build(std::string NVCC_FLAGS) {
-		HPX_ASSERT(this->get_gid());
-		typedef server::program::build_action action_type;
-		return hpx::async < action_type > (this->get_gid(),NVCC_FLAGS);
-	}
+    hpx::lcos::future<void> build(std::string NVCC_FLAGS) {
+        HPX_ASSERT(this->get_gid());
+        typedef server::program::build_action action_type;
+        return hpx::async < action_type > (this->get_gid(),NVCC_FLAGS);
+    }
 
-	void build_sync(std::string NVCC_FLAGS) {
-		HPX_ASSERT(this->get_gid());
-		build(NVCC_FLAGS).get();
-	}
+    void build_sync(std::string NVCC_FLAGS) {
+        HPX_ASSERT(this->get_gid());
+        build(NVCC_FLAGS).get();
+    }
 
-	hpx::lcos::future<void> create_kernel(std::string module_name,
-			std::string kernel_name) {
-		HPX_ASSERT(this->get_gid());
-		typedef server::program::create_kernel_action action_type;
-		hpx::async<action_type>(this->get_gid(),module_name,kernel_name);
-	}
+    hpx::lcos::future<void> create_kernel(std::string module_name,
+            std::string kernel_name) {
+        HPX_ASSERT(this->get_gid());
+        typedef server::program::create_kernel_action action_type;
+        hpx::async<action_type>(this->get_gid(),module_name,kernel_name);
+    }
 
-	void create_kernel_sync(std::string module_name, std::string kernel_name) {
+    void create_kernel_sync(std::string module_name, std::string kernel_name) {
 
-		create_kernel(module_name, kernel_name).get();
-	}
+        create_kernel(module_name, kernel_name).get();
+    }
 
-	void set_source_sync(std::string source) {
-		HPX_ASSERT(this->get_gid());
-		 typedef server::program::set_source_action action_type;
-		 hpx::async<action_type>(this->get_gid(),source).get();
-	}
+    void set_source_sync(std::string source) {
+        HPX_ASSERT(this->get_gid());
+         typedef server::program::set_source_action action_type;
+         hpx::async<action_type>(this->get_gid(),source).get();
+    }
 };
 }
 }
