+++ cuda/server/get_devices.hpp
@@ -6,17 +6,16 @@
 #ifndef HPX_CUDA_SERVER_GET_DEVICES_HPP_
 #define HPX_CUDA_SERVER_GET_DEVICES_HPP_
 
-#include <hpx/hpx.hpp>
-#include <hpx/config.hpp>
-#include <hpx/runtime/actions/component_action.hpp>
-#include <hpx/runtime/components/server/managed_component_base.hpp>
-#include <hpx/runtime/components/server/locking_hook.hpp>
+#include <hpx/include/components.hpp>
+
+#include "cuda/fwd_declarations.hpp"
+#include "cuda.hpp"
+#include "cuda/cuda_error_handling.hpp"
+#include "cuda/export_definitions.hpp"
 
 #include <cuda.h>
-#include "../../cuda.hpp"
-#include "../CudaErrorHandling.hpp"
 
-#include "../fwd_declarations.hpp"
+#include <vector>
 
 ////////////////////////////////////////////////////////////////
 namespace hpx {
@@ -25,15 +24,11 @@ namespace server {
 
 // /////////////////////////////////////////////////////
 //  Global cuda functions
-//
 
 // Returns the IDs of all devices on current host
-std::vector<hpx::cuda::device>
-get_devices(int major, int minor);
+HPX_CUDA_EXPORT std::vector<hpx::cuda::device> get_devices(int major, int minor);
 
-//[opencl_management_action_types
 HPX_DEFINE_PLAIN_ACTION(get_devices, get_devices_action);
-//]
 
 }
 }
