+++ cuda/server/program_server.cpp
@@ -1,26 +1,21 @@
-// Copyright (c)		2013 Damond Howard
-//						2015 Patrick Diehl
+// Copyright (c)    2013 Damond Howard
+//                  2015 Patrick Diehl
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt
 
-//#include <hpx/hpx_fwd.hpp>
-//#include <hpx/runtime/components/server/managed_component_base.hpp>
-//#include <hpx/runtime/components/server/locking_hook.hpp>
-//#include <hpx/runtime/actions/component_action.hpp>
-//#include <hpx/include/util.hpp>
-//#include <hpx/hpx_init.hpp>
-//#include <hpx/include/runtime.hpp>
+#include <hpx/hpx.hpp>
 
-//#include <sstream>
-//#include <cstring>
-//#include <cuda.h>
-//#include <cuda_runtime.h>
+#include <cuda.h>
+#include <cuda_runtime.h>
 
-#include "program.hpp"
-//#include "kernel.hpp"
-//#include "../kernel.hpp"
+#include "cuda/server/program.hpp"
+#include "cuda/kernel.hpp"
 
-using namespace hpx::cuda::server;
+#include <sstream>
+#include <string>
+
+namespace hpx { namespace cuda { namespace server
+{
 
 program::program(){}
 
@@ -88,4 +83,6 @@ void program::build(std::string NVCC_FLAGS)
 	//	exit(1);
 	//}
 }
+
+}}}
