--- cuda/server/std.hpp
@@ -1,5 +1,0 @@
-// Copyright (c)    2014 Damond Howard
-//
-// Distributed under the Boost Software License, Version 1.0. (See accompanying
-// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-
